Redo of Binary and the start of implementing Boolean
For Boolean I'm thinking it'll be easier to just find a way to
have each operand (blue, outline, solid etc.) keep a bitstring
or something of the indexes of the objects that they would hide.
This way the operators should just be able to && and || them
together to evaluate the boolean statements. Parsing the statements
should be as easy as starting at the left and keeping up with the
last result(the bitstring of indexes to hide).

Haven't done much with the DnD behavior for Boolean, want it to
eventually enforce rules of making sure there is a operator between
two properties and I need to figure out how to handle grouping, I guess
that would just be a source that is also draggable but that handles
stacking nicely.